refactor: BridgeMaker, Bridge 리팩터링
Bridge객체에서 다리 사이즈 유효성을 검사하느것보다 BridgeMaker에서 검사하는것이 해당 객체에 맞는 기능

RF-1
bridgeSize를 BridgeMaker에서 유효성검사하도록 벽녕
Bridge객체는 List<String> bridge만을 가지도록 구성
