+++ src/test/java/bridge/BridgeMakerTest.java
@@ -9,6 +9,9 @@ import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
+import util.ErrorMessage;
 
 class BridgeMakerTest {
     private final BridgeRandomNumberGenerator bridgeRandomNumberGenerator = new BridgeRandomNumberGenerator();
@@ -35,5 +38,27 @@ class BridgeMakerTest {
         //then
         assertThat(bridge).containsAnyOf(UP_STAIRS,DOWN_STAIRS);
 
+    }
+
+
+    @DisplayName("[Bridge] 다리 길이의 입력값이 범위내 없는경우 예외발생")
+    @ParameterizedTest
+    @ValueSource(ints = {2, -1, 132, 21})
+    void notCorrectBridgeSize(int size) {
+        //given
+        //when
+        //then
+        assertThatThrownBy(() -> BridgeMaker.validateBridgeSize(size))
+                .isInstanceOf(IllegalArgumentException.class).hasMessage(ErrorMessage.NOT_RANGE_BRIDGE_SIZE);
+    }
+
+    @DisplayName("[Bridge] 다리 길이의 입력값이 정상 범위인경우")
+    @ParameterizedTest
+    @ValueSource(ints = {3, 14, 20})
+    void correctBridgeSize(int size) {
+        //given
+        //when
+        //then
+        assertDoesNotThrow(() -> BridgeMaker.validateBridgeSize(size));
     }
 }
